requirements.txt
psycopg2-binary
python-dotenv
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2
gunicorn==21.2.0
requests==2.32.5
